Vector-ICs - speed towards the monomorphic exit as quickly as possible.
Thanks to some careful assumptions, we can examine the object found at
vector[slot] and trust it's a heap object where the second field is
either a map if it's a WeakCell, or definitely not a map if it's a
Symbol, String or FixedArray. Use this to save a memory read.

BUG=

Review URL: https://codereview.chromium.org/1032163002

Cr-Commit-Position: refs/heads/master@{#27631}