src: Changed branch in MacroAssembler::CheckPageFlag to always assume far branch target
Assuming near branches in the bne command in MacroAssembler::CheckPageFlag can
result in problems when a far branch is desired, as then the branch target
will be inserted incorrectly into the assembly code in such a way that the
target will be executed as an instruction. This causes a crash.

This problem has been resolved by always assuming a far branch. While this
does cause a tradeoff in code size, the tradeoff should be minimal
(several kb at most).