[z/OS] Fix __sync_val_compare_and_swap return.
The old implementation could return the incorrect value if *ptr was
modified between assigning tmp and performing __cs1.
The new implementation takes advantage of the fact that the cs
instruction on s390 will update oldval to *ptr when the swap is not
successful so that after cs1 oldval will always have the value that was
in ptr before cs1 executed.