[zOS] Fix callee-save reg for handleScope->Next
In CallApiFunctionAndReturn, it saves the current handle scope data
in a set of callee-saved registers.  There is a bug on zOS where the
current handleScope->next is saved in r6, which is not a preserved
register under XPLINK.  R6 is the entry point of a called routine,
and in certain cases if the callee does not restore the original R6,
we end up incorrectly storing the wrong value when restoring handleScope->next.