[z/OS] Use custom semaphore implementation
BSD semaphores need to be explicitly deallocated by
a process for the operating system to reclaim it on
USS. This leaves stale semaphores around whenever a
v8 based application terminates abnormally.
To address this issue we are replacing
the use of BSD semaphores with custom semaphore
implementation which uses pthread mutexes/condition
variables. Any pthread mutexes/condition variables
that are left behind by a terminated process are
cleaned up by the operating system.