[z/OS] Enable RAS features
All the literals in the v8 source are converted
to ascii during compilation. This is needed for
functional correctness(Javascript is UTF-8 based)
However this leads to all trace/debug content from
--print-all-code, --trace and other debug options
to be mangled when its printed on stdout/stderr.

To fix this we create a wrapper class for std::ostream
and std::ofstream and overload their output operator
to convert the streams on the fly.